Exception handling - delete

+++ src/main/java/com/shiringu/course/services/UserService.java
@@ -4,21 +4,24 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.shiringu.course.entities.User;
 import com.shiringu.course.repositories.UserRepository;
+import com.shiringu.course.services.exceptions.DatabaseException;
 import com.shiringu.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long Id) {
 		Optional<User> obj = repository.findById(Id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(Id));
@@ -27,14 +30,23 @@ public class UserService {
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (repository.existsById(id)) 
+				repository.deleteById(id);
+			else 
+				throw new ResourceNotFoundException(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
-		updateData(entity,obj);
+		updateData(entity, obj);
 		return repository.save(entity);
 	}
